open|creat: skip S_ISGID check on files created by non-group members
0fa3ecd87848 ("Fix up non-directory creation in SGID directories")
fixes problem described in CVE-2018-13405. This commit is getting
backported to older streams as well.

This patch removes S_ISGID check for files created by non-group members
in LTP tests creat08 and open10.

Once 0fa3ecd87848 will be in non-rc kernel, we could add a new test
for this CVE that would be limited to 4.18+ kernels.

Resolves: LTP-1
Signed-off-by: Anson <aouyang@sierrawireless.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Naresh Kamboju <naresh.kamboju@linaro.org>
Change-Id: I9287c97e970e0b738426c0c227df8c13ae7d755e